LayoutTests/http/tests/in-app-browser-privacy/app-bound-attribution-preflight-async-expected.txt

CONSOLE MESSAGE: Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/status-404-without-body.py due to access control checks.
Tests app bound requests get marked for async preflights

PASS successfully loaded only app-bound requests
